Restricting access to exhibits

+++ src/main/java/pl/wszeborowski/mateusz/exhibit/ExhibitService.java
@@ -4,16 +4,20 @@ import lombok.NoArgsConstructor;
 import pl.wszeborowski.mateusz.exhibit.model.Exhibit;
 import pl.wszeborowski.mateusz.museum.MuseumService;
 import pl.wszeborowski.mateusz.museum.model.Museum;
+import pl.wszeborowski.mateusz.user.User;
 
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
+import javax.servlet.http.HttpServletRequest;
 import javax.transaction.Transactional;
+import java.security.AccessControlException;
 import java.util.Collection;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
+
 
 @ApplicationScoped
 @NoArgsConstructor
@@ -23,6 +27,9 @@ public class ExhibitService {
     private EntityManager em;
 
     private MuseumService museumService;
+
+    @Inject
+    private HttpServletRequest securityContext;
 
     @Inject
     public ExhibitService(MuseumService museumService) {
@@ -61,15 +68,23 @@ public class ExhibitService {
 
     @Transactional
     public synchronized void saveExhibit(Exhibit exhibit) {
-        if (exhibit.getId() == null) {
-            em.persist(exhibit);
-        } else {
-            em.merge(exhibit);
+        if (securityContext.isUserInRole(User.Roles.USER)) {
+            if (exhibit.getId() == null) {
+                em.persist(exhibit);
+            } else {
+                em.merge(exhibit);
+            }
+            return;
         }
+        throw new AccessControlException("Access denied");
     }
 
     @Transactional
     public void removeExhibit(Exhibit exhibit) {
-        em.remove(em.merge(exhibit));
+        if (securityContext.isUserInRole(User.Roles.USER)) {
+            em.remove(em.merge(exhibit));
+            return;
+        }
+        throw new AccessControlException("Access denied");
     }
 }
